Added Amount, tests need fixing

+++ shoppingcart/src/main/kotlin/es/katas/repositories/ProductRepository.kt
@@ -3,12 +3,12 @@ package es.katas.repositories
 import es.katas.domain.Product
 
 class ProductRepository {
-    private val products = mutableMapOf<Int, Product>()
+    private val products = mutableMapOf<Long, Product>()
 
-    fun addProduct(productId: Int, product: Product): ProductRepository {
+    fun addProduct(productId: Long, product: Product): ProductRepository {
         products[productId] = product
         return this
     }
 
-    fun getProduct(productId: Int) = products[productId]
+    fun getProduct(productId: Long) = products[productId]
 }
